Make requestOAuthAccesToken open
We have an interesting use case with Microsoft's OAuth (Azure AD, Graph API, Outlook API, etc)

Oauth tokens are scoped to a single resource. However, Microsoft allows refresh tokens to be used to request additional tokens for other resources.  The easiest way to do this is to call requestOAuthAccessToken for each additional resource we intend to use with the additional resource parameters we need to specify.

Opening requestOAuthAccessToken makes it much simpler for us to request additional access tokens.