If the client_secret is empty, don't post it at all.
Really we shouldn't be storing a client secret in an app at all, we should be using
PKCE. Some servers will reject the request if it's a "public" app and you include
the secret. Since it would be tricky to fully make it optional in the client everywhere,
I'm changing it to just ignore empty strings.