Fix unowned self
With an unowned self the reference to self can become nil without the
ability for the code to be aware. As such, unowned self should only be
used when we know something else is guaranteed to retain a reference to
our self. In this case there is no such guarantee. Fortunately, we can
use a weak self to have nullability safety here without any performance
concerns.